requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.1
alembic==1.15.2
altair==5.1.2
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==3.0.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.3.0
auth0-python==4.9.0
azure-ai-formrecognizer==3.3.0
azure-ai-language-questionanswering==1.1.0
azure-common==1.1.28
azure-core==1.29.4
azure-identity==1.14.0
azure-nspkg==3.0.2
azure-search-documents==11.4.0b8
azure-storage-blob==12.18.3
azure-storage-file-datalake==12.13.1
babel==2.17.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.1
beautifulsoup4==4.13.3
billiard==4.2.0
bleach==6.2.0
blinker==1.7.0
boto3==1.26.122
botocore==1.29.122
build==1.2.2.post1
cachetools==5.3.2
captcha==0.6.0
celery==5.3.1
certifi==2024.12.14
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudevents==1.11.0
cohere==5.14.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confuse==2.0.1
contourpy==1.3.1
crewai==0.86.0
crewai-tools==0.17.0
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.1
debugpy==1.8.12
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
desert==2022.9.22
distlib==0.3.7
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
docstring_parser==0.16
docx2txt==0.9
durationpy==0.9
embedchain==0.1.128
et-xmlfile==1.1.0
executing==2.2.0
extra-streamlit-components==0.1.71
fastapi==0.115.9
fastavro==1.10.0
fastjsonschema==2.20.0
filelock==3.13.1
Flask==2.3.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.3
flatbuffers==23.5.26
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.12.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
googleapis-common-protos==1.69.2
gptcache==0.1.44
greenlet==3.0.1
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.1
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.19.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.1.0
instructor==1.7.9
invoke==2.2.0
ipykernel==6.29.5
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.6
jinja2-time==0.2.0
jiter==0.8.2
jmespath==1.0.1
joblib==1.3.2
json5==0.10.0
json_repair==0.40.0
jsonpatch==1.33
jsonpickle==4.0.5
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2023.7.1
jupyter==1.1.1
jupyter-client==6.1.12
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==1.24.0
jupyter_core==5.7.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
kombu==5.3.5
kubernetes==32.0.1
lancedb==0.21.2
langchain==0.3.17
langchain-cohere==0.3.5
langchain-community==0.3.16
langchain-core==0.3.33
langchain-experimental==0.3.4
langchain-openai==0.2.12
langchain-text-splitters==0.3.5
langfuse==1.0.18
langgraph==0.4.3
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.69
langsmith==0.3.27
lazy-object-proxy==1.10.0
litellm==1.60.0
load-dotenv==0.1.0
loguru==0.7.2
make==0.1.6.post2
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.1.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.86
mistune==3.1.2
mmh3==4.0.1
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
msal==1.24.1
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.2.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.8.1
nodeenv==1.9.1
notebook==6.5.5
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.7
onnxruntime==1.16.3
openai==1.57.4
openapi-core==0.19.4
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.16
ormsgpack==1.9.1
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.2.3
pandas-stubs==2.1.1.230928
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.4
pathable==0.4.3
pdfminer.six==20250327
pdfplumber==0.11.6
pickleshare==0.7.5
Pillow==10.1.0
pinecone-client==2.2.2
pipenv==2024.0.1
platformdirs==3.11.0
pluggy==1.5.0
poetry-core==1.9.0
portalocker==2.8.2
posthog==3.23.0
prance==23.6.21.0
prometheus_client==0.21.1
prompt-toolkit==3.0.43
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pulsar-client==3.3.0
pure_eval==0.2.3
pyarrow==14.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybars4==0.9.13
pycparser==2.21
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.10.1
PyMeta3==0.5.1
PyMySQL==1.1.1
pynessie==0.67.0
pyparsing==3.2.1
pypdf==5.4.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pyright==1.1.398
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==3.2.1
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.15
PyYAML==6.0.1
pyzmq==24.0.1
qdrant-client==1.13.3
rapidfuzz==3.9.6
redis==5.0.0
referencing==0.30.2
regex==2024.11.6
requests==2.32.3
requests-aws4auth==1.3.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.12.0
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.8
s3transfer==0.6.2
schema==0.7.7
scikit-learn==1.6.0
scipy==1.15.2
selenium==4.31.0
semantic-kernel==1.21.3
Send2Trash==1.8.3
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.45.3
streamlit==1.41.1
streamlit-authenticator==0.4.1
streamlit-chat==0.1.1
sympy==1.12
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.15.0
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
tomlkit==0.13.0
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
trove-classifiers==2024.7.2
typer==0.9.0
types-python-dateutil==2.8.19.14
types-pytz==2023.3.1.1
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uri-template==1.3.0
urllib3==2.3.0
uuid==1.30
uv==0.6.13
uvicorn==0.24.0.post1
validators==0.22.0
vine==5.1.0
virtualenv==20.24.5
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==2.3.1
widgetsnbextension==4.0.13
win32-setctime==1.1.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.19.0
zipp==3.17.0
zstandard==0.23.0
